Fixed teamserver routes
The openvpn-server (seeds) push route to all clients to access the
target network (192.168.42.0/24). But the teamserver (forbidden) is
already connected to this network and its the gateway for this
new route... So, when the openvpn server push routes to clients, the
teamserver get a second route for the target network.. which is
unreachable.

The solution here is to override the openvpn's client template to add
extras configuration, forbidden needs to ignore routes from the
vpn server and configure itself the route for the vpn network.
Extra configuration is take from node[openvpn][config][extras]
attribute, it is a Hash of keys/values containing extra configuration.